Combine Tool and McpAIFunction to McpFunction
When we cannot predict the specific schema returned by the server, the existing predefined JsonSchema is likely to be invalid, so it is replaced with JsonElement. When InputSchema becomes JsonElement, the definition of Tool is basically the same as AIFunction, so the two can be further merged. After that, we no longer need to worry about the conversion between Tool and AIFunction.